+++ hyperutil-reference/src/main/java/xyz/ielis/hyperutil/reference/fasta/SequenceInterval.java
@@ -2,13 +2,56 @@ package xyz.ielis.hyperutil.reference.fasta;
 
 import de.charite.compbio.jannovar.reference.GenomeInterval;
 
+import java.util.HashMap;
+import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.regex.Pattern;
 
+/**
+ * This class puts together a nucleotide sequence and a particular {@link GenomeInterval}. Each instance is built
+ * using the builder while checks are performed at the end of the building process.
+ */
 public class SequenceInterval {
 
-    private static final Pattern BASES = Pattern.compile("[ACGTNacgtn]+");
+    /**
+     * Map for translation of a single nucleotide/base symbol into its reverse complement base
+     */
+    private static final Map<Character, Character> IUPAC_COMPLEMENT_MAP;
+
+    static {
+        Map<Character, Character> TEMPORARY = new HashMap<>();
+        TEMPORARY.putAll(
+                Map.of(
+                        // STANDARD
+                        'A', 'T',
+                        'C', 'G',
+                        'G', 'C',
+                        'T', 'A',
+                        'U', 'A'));
+        TEMPORARY.putAll(
+                Map.of(
+                        // AMBIGUITY BASES - first part
+                        'W', 'W', // weak - A,T
+                        'S', 'S', // strong - C,G
+                        'M', 'K', // amino - A,C
+                        'K', 'M', // keto - G,T
+                        'R', 'Y', // purine - A,G
+                        'Y', 'R')); // pyrimidine - C,T
+
+        TEMPORARY.putAll(
+                Map.of(
+                        // AMBIGUITY BASES - second part
+                        'B', 'V', // not A
+                        'D', 'H', // not C
+                        'H', 'D', // not G
+                        'V', 'B', // not T
+                        'N', 'N' // any one base
+                )
+        );
+
+        IUPAC_COMPLEMENT_MAP = Map.copyOf(TEMPORARY);
+    }
+
 
     private final GenomeInterval interval;
     private final String sequence;
@@ -20,9 +63,6 @@ public class SequenceInterval {
             throw new IllegalArgumentException(String.format("Lengths do not match: interval %s != sequence %s",
                     interval.length(), sequence.length()));
         }
-        if (!BASES.matcher(sequence).matches()) {
-            throw new IllegalArgumentException("Sequence does not match regexp `[ACGTNacgtn]+`");
-        }
     }
 
     public static Builder builder() {
@@ -30,40 +70,30 @@ public class SequenceInterval {
     }
 
     /**
-     * Convert nucleotide sequence to reverse complement.
+     * Convert nucleotide sequence to reverse complement. If <code>U</code> is present, it is complemented to
+     * <code>A</code> as expected. However, <code>A</code> is always complemented to <code>T</code>.
      *
-     * @param sequence of nucleotides, only {a,c,g,t,n,A,C,G,T,N} permitted
+     * @param sequence of nucleotides in IUPAC notation
      * @return reverse complement of given <code>sequence</code>
-     * @throws IllegalArgumentException if there is an unpermitted character present
      */
     static String reverseComplement(String sequence) {
         char[] oldSeq = sequence.toCharArray();
         char[] newSeq = new char[oldSeq.length];
         int idx = oldSeq.length - 1;
         for (int i = 0; i < oldSeq.length; i++) {
-            if (oldSeq[i] == 'A') {
-                newSeq[idx - i] = 'T';
-            } else if (oldSeq[i] == 'a') {
-                newSeq[idx - i] = 't';
-            } else if (oldSeq[i] == 'T') {
-                newSeq[idx - i] = 'A';
-            } else if (oldSeq[i] == 't') {
-                newSeq[idx - i] = 'a';
-            } else if (oldSeq[i] == 'C') {
-                newSeq[idx - i] = 'G';
-            } else if (oldSeq[i] == 'c') {
-                newSeq[idx - i] = 'g';
-            } else if (oldSeq[i] == 'G') {
-                newSeq[idx - i] = 'C';
-            } else if (oldSeq[i] == 'g') {
-                newSeq[idx - i] = 'c';
-            } else if (oldSeq[i] == 'N') {
-                newSeq[idx - i] = 'N';
-            } else if (oldSeq[i] == 'n') {
-                newSeq[idx - i] = 'n';
-            } else
-                throw new IllegalArgumentException(String.format("Illegal nucleotide %s in sequence %s",
-                        oldSeq[i], sequence));
+            char template = oldSeq[i];
+            boolean isUpperCase = Character.isUpperCase(template);
+
+            char toLookUp = isUpperCase
+                    ? template // no-op, the IUPAC map contains upper-case characters
+                    : Character.toUpperCase(template);
+
+            char complement = IUPAC_COMPLEMENT_MAP.getOrDefault(toLookUp, template);
+            char complementWithCase = isUpperCase
+                    ? complement // no-op, the IUPAC map contains upper-case characters
+                    : Character.toLowerCase(complement);
+
+            newSeq[idx - i] = complementWithCase;
         }
         return new String(newSeq);
     }
@@ -134,6 +164,11 @@ public class SequenceInterval {
             return this;
         }
 
+        /**
+         * @return built {@link SequenceInterval}
+         * @throws IllegalArgumentException if the {@link #sequence} length does not match length of the provided
+         *                                  {@link #interval}
+         */
         public SequenceInterval build() {
             return new SequenceInterval(this);
         }